Switch to non-encrypted storage
