+++ dependencies.gradle
@@ -23,7 +23,6 @@ ext.versions = [
         androidxNavigation          : '2.5.3',
         androidxPreference          : '1.2.0',
         androidxRecyclerview        : '1.2.1',
-        androidxSecurity            : '1.1.0-alpha05',
         binaryCompatibilityValidator: '0.13.1',
         cameraX                     : '1.3.1',
         dagger                      : '2.48',
@@ -118,7 +117,6 @@ ext.libs = [
                 navigationUi      : "androidx.navigation:navigation-ui-ktx:${versions.androidxNavigation}",
                 preference        : "androidx.preference:preference-ktx:${versions.androidxPreference}",
                 recyclerView      : "androidx.recyclerview:recyclerview:${versions.androidxRecyclerview}",
-                security          : "androidx.security:security-crypto:${versions.androidxSecurity}",
         ],
         camera                              : [
                 core     : "androidx.camera:camera-core:${versions.cameraX}",
