+++ link/src/main/java/com/stripe/android/link/LinkPaymentLauncher.kt
@@ -4,7 +4,7 @@ import androidx.activity.result.ActivityResultCaller
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.ActivityResultRegistry
 import androidx.annotation.RestrictTo
-import com.stripe.android.link.account.CookieStore
+import com.stripe.android.link.account.LinkStore
 import com.stripe.android.link.injection.LinkAnalyticsComponent
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -17,7 +17,7 @@ import javax.inject.Singleton
 class LinkPaymentLauncher @Inject internal constructor(
     linkAnalyticsComponentBuilder: LinkAnalyticsComponent.Builder,
     private val linkActivityContract: LinkActivityContract,
-    private val cookieStore: CookieStore,
+    private val linkStore: LinkStore,
 ) {
     private val analyticsHelper = linkAnalyticsComponentBuilder.build().linkAnalyticsHelper
 
@@ -42,7 +42,7 @@ class LinkPaymentLauncher @Inject internal constructor(
         linkActivityResultLauncher = activityResultCaller.registerForActivityResult(
             linkActivityContract
         ) { linkActivityResult ->
-            cookieStore.markLinkAsUsed()
+            linkStore.markLinkAsUsed()
             analyticsHelper.onLinkResult(linkActivityResult)
             callback(linkActivityResult)
         }
