+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/LinkHandler.kt
@@ -7,7 +7,7 @@ import com.stripe.android.link.LinkConfiguration
 import com.stripe.android.link.LinkConfigurationCoordinator
 import com.stripe.android.link.LinkPaymentDetails
 import com.stripe.android.link.LinkPaymentLauncher
-import com.stripe.android.link.account.CookieStore
+import com.stripe.android.link.account.LinkStore
 import com.stripe.android.link.analytics.LinkAnalyticsHelper
 import com.stripe.android.link.injection.LinkAnalyticsComponent
 import com.stripe.android.link.model.AccountStatus
@@ -35,7 +35,7 @@ internal class LinkHandler @Inject constructor(
     private val linkLauncher: LinkPaymentLauncher,
     private val linkConfigurationCoordinator: LinkConfigurationCoordinator,
     private val savedStateHandle: SavedStateHandle,
-    private val cookieStore: CookieStore,
+    private val linkStore: LinkStore,
     linkAnalyticsComponentBuilder: LinkAnalyticsComponent.Builder,
 ) {
     sealed class ProcessingState {
@@ -177,7 +177,7 @@ internal class LinkHandler @Inject constructor(
             ).getOrNull()
 
             if (linkPaymentDetails != null) {
-                cookieStore.markLinkAsUsed()
+                linkStore.markLinkAsUsed()
             }
 
             _processingState.emit(
